Migrate the 'winclang-format' hook to invoke
